requirements.txt
pyspark==3.5.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.4
schedule==1.2.1
